fix(webdriver): rename files in CamelCase style
