+++ webdriver-ts/package.json
@@ -7,12 +7,12 @@
   "scripts": {
     "tsbench": "cross-env LANG=\"en_US.UTF-8\" ts-node --files src/benchmarkRunner.ts",
     "bench": "cross-env LANG=\"en_US.UTF-8\" node dist/benchmarkRunner.js",
-    "checkCSP": "cross-env LANG=\"en_US.UTF-8\" node dist/isCSPCompliant.js",
+    "checkCSP": "cross-env LANG=\"en_US.UTF-8\" node dist/isCspCompliant.js",
     "compile": "tsc",
-    "results": "node dist/createResultJS.js && cd .. && cd webdriver-ts-results && npm run build",
-    "resultsChrome": "node dist/createResultJS.js --browser chrome && cd .. && cd webdriver-ts-results && npm run build-prod",
-    "resultsFirefox": "node dist/createResultJS.js --browser firefox && cd .. && cd webdriver-ts-results && npm run build-prod",
-    "resultsSafari": "node dist/createResultJS.js --browser safari && cd .. && cd webdriver-ts-results && npm run build-prod",
+    "results": "node dist/createResultJs.js && cd .. && cd webdriver-ts-results && npm run build",
+    "resultsChrome": "node dist/createResultJs.js --browser chrome && cd .. && cd webdriver-ts-results && npm run build-prod",
+    "resultsFirefox": "node dist/createResultJs.js --browser firefox && cd .. && cd webdriver-ts-results && npm run build-prod",
+    "resultsSafari": "node dist/createResultJs.js --browser safari && cd .. && cd webdriver-ts-results && npm run build-prod",
     "isKeyed": "cross-env LANG=\"en_US.UTF-8\" node dist/isKeyed.js",
     "test": "vitest --run",
     "test:watch": "vitest",
