+++ webdriver-ts/src/createResultJs.ts
